# contracts/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "WorkBuddy Buddy Control Plane API",
    "version": "1.0.0",
    "summary": "Versioned Edge reporting and public Office viewing contract.",
    "description": "Milestone 0 contract for the WorkBuddy Buddy Control Plane.\n\nEdge reports are authenticated, signed, and fenced by `instance_id`, a\ncompare-and-swap-linked `boot_id`, and the per-boot `sequence`. A new boot is\naccepted only when its required `previous_boot_id` names the server's current\nboot (or is null for the first boot). State events and heartbeats share the\nsame strictly increasing, per-boot sequence space.\n\nThe server assigns one authoritative `server_received_at` to every accepted report. Client\n`observed_at` is diagnostic metadata only and MUST NOT determine leases,\ngrace periods, activity intervals, rankings, or awards. Likewise,\n`idle_stage` is never accepted as a scoring input: public idle stages are\nderived by the server from continuous eligible-idle time.\n\nProtocol v1 publishes `current_protocol = 1` and\n`min_supported_protocol = 1`; it does not imply or accept a fictional v0.\nWhen protocol N is introduced, the same response fields can advertise the\nN/N-1 compatibility window. Unsupported versions receive a structured `426`\nresponse so an Edge can present an actionable upgrade message without\ndiscarding its local desktop-pet state.\n",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
...
        }
...
  "paths": {
    "/api/v1/edge/events:batch": {
      "post": {
        "tags": [
          "Edge"
        ],
        "summary": "Report a bounded batch of derived state events.",
        "description": "Durably and atomically accepts an ordered, contiguous batch of\nderived Edge events from one boot's monotonic sequence space. State\ntransitions and heartbeats consume that same space. No\nprefix of a rejected batch is applied. While that boot remains current,\nreplaying an entire batch with the same `instance_id`, accepted boot\ngeneration, sequence range, and canonical payload returns the same\nacknowledgement without side effects. A replay from a fenced boot always\nreturns `stale_boot` instead. Reusing any sequence with different canonical\ncontent is a conflict. A gap or partially overlapping replay is also a\nconflict; callers recover from the `expected_sequence` in the Problem response\nrather than skipping ahead.\n\nA new boot starts at sequence 1 and its first event MUST be a complete\n`state_transition` snapshot. Its required `previous_boot_id` is null only for\nthe first boot; otherwise it MUST exactly match the server's current boot.\nThis compare-and-swap establishes the next fencing generation without allowing\na delayed, previously unseen old boot to reverse-fence the current one. A\nmismatch or report from a fenced boot is rejected and MUST NOT update current\npresence, renew a lease, restart a grace period, or contribute to scoring.\n\nOnly privacy-safe derived state may be sent. Prompt text, messages, tool\narguments, filesystem paths, session contents, email bodies, OAuth\ntokens, raw spool records, and client-computed `idle_stage` are outside\nthis contract.\n",
        "operationId": "reportEdgeEventBatchV1",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/edge-report-envelope.v1.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/EdgeReportAccepted"
          },
          "400": {
            "$ref": "#/components/responses/BadEdgeReport"
          },
          "401": {
            "$ref": "#/components/responses/UnauthenticatedEdge"
          },
          "409": {
            "$ref": "#/components/responses/EdgeFenceConflict"
          },
          "413": {
            "$ref": "#/components/responses/EdgeBatchTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEdgeReport"
          },
          "426": {
            "$ref": "#/components/responses/UnsupportedProtocol"
          },
          "429": {
            "$ref": "#/components/responses/EdgeRateLimited"
          },
          "503": {
            "$ref": "#/components/responses/TemporarilyUnavailable"
          }
        }
      }
    },
    "/api/v1/edge/heartbeat": {
      "post": {
        "tags": [
          "Edge"
        ],
        "summary": "Renew presence with a sequenced heartbeat.",
        "description": "Accepts an Edge report envelope containing exactly one `heartbeat`\nevent for the already-established current boot. A heartbeat cannot establish\na new boot: sequence 1 of every new boot is sent to the batch endpoint as a\ncomplete state snapshot. Heartbeats consume the same monotonic sequence space\nas state transitions; they are not an independent counter. Replaying the same\ncanonical heartbeat returns the same acknowledgement without renewing the\nlease twice. A gap, a partial overlap, a multi-event envelope, or a\nnon-heartbeat event is rejected.\n\nA heartbeat may renew an active lease, but it cannot backfill an expired\nlease or recover idle time lost while offline. Lease calculations begin\nfrom the server's `server_received_at`, never from `observed_at` or `sent_at`.\nA heartbeat from a fenced boot is incapable of changing presence.\n",
        "operationId": "reportEdgeHeartbeatV1",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/edge-report-envelope.v1.schema.json#/$defs/heartbeatEnvelope"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/EdgeReportAccepted"
          },
          "400": {
            "$ref": "#/components/responses/BadEdgeReport"
          },
          "401": {
            "$ref": "#/components/responses/UnauthenticatedEdge"
          },
          "409": {
            "$ref": "#/components/responses/EdgeFenceConflict"
          },
          "413": {
            "$ref": "#/components/responses/EdgeBatchTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEdgeReport"
          },
          "426": {
            "$ref": "#/components/responses/UnsupportedProtocol"
          },
          "429": {
            "$ref": "#/components/responses/EdgeRateLimited"
          },
          "503": {
            "$ref": "#/components/responses/TemporarilyUnavailable"
          }
        }
      }
    },
    "/api/v1/onboarding/offices": {
      "post": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Create a private Office and issue one pairing.",
        "description": "Atomically creates a private, unlisted Office with one owner, Lobby, logical Agent, mounted pet, weekday schedule, and short-lived Edge pairing. The raw `pairing_code`, `status_token`, and public Office capability are returned only once and are stored by the server only as hashes. This anonymous convenience endpoint is IP-rate-limited.",
        "operationId": "createOnboardingOfficeV1",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfficeOnboardingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/OnboardingOfficeCreated"
          },
          "400": {
            "$ref": "#/components/responses/OnboardingBadRequest"
          },
          "429": {
            "$ref": "#/components/responses/OnboardingRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/OnboardingInternalError"
          }
        }
      }
    },
    "/api/v1/onboarding/pairings/{status_token}": {
      "get": {
        "tags": [
          "Onboarding"
        ],
        "summary": "Read the status of one short-lived pairing.",
        "description": "Returns only `pending`, `claimed`, or `expired` for the supplied status capability. The 43-character status token is distinct from both the Edge pairing code and the public Office token. Unknown capabilities return `404` without revealing Office existence.",
        "operationId": "getOnboardingPairingStatusV1",
        "parameters": [
          {
            "$ref": "#/components/parameters/PairingStatusToken"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OnboardingPairingStatus"
          },
          "400": {
            "$ref": "#/components/responses/OnboardingBadRequest"
          },
          "404": {
            "$ref": "#/components/responses/OnboardingPairingNotFound"
          },
          "429": {
            "$ref": "#/components/responses/OnboardingRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/OnboardingInternalError"
          }
        }
      }
    },
    "/api/v1/edge/enrollment/claim": {
      "post": {
        "tags": [
          "Onboarding",
          "Edge"
        ],
        "summary": "Claim a pairing with an Ed25519 device key.",
        "description": "Consumes a pending pairing and enrolls one active Agent instance with the supplied raw 32-byte Ed25519 public key. The key is encoded as canonical padded standard Base64. A replay using the same pairing and public key returns the exact same enrollment result; a different key returns `409`. Expired pairings return `410`. The private key never leaves WorkBuddy.",
        "operationId": "claimEdgeEnrollmentPairingV1",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimPairingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/EdgeEnrollmentClaimed"
          },
          "400": {
            "$ref": "#/components/responses/OnboardingBadRequest"
          },
          "404": {
            "$ref": "#/components/responses/OnboardingPairingNotFound"
          },
          "409": {
            "$ref": "#/components/responses/OnboardingPairingConflict"
          },
          "410": {
            "$ref": "#/components/responses/OnboardingPairingExpired"
          },
          "429": {
            "$ref": "#/components/responses/OnboardingRateLimited"
          },
          "500": {
            "$ref": "#/components/responses/OnboardingInternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "Optional request-attempt correlation key. Edge event idempotency does\nnot depend on this header; it is always enforced by\n`instance_id + accepted boot generation + sequence`.\n",
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._~-]+$"
        }
      },
...
    },
...
    "schemas": {
      "CreateOfficeOnboardingRequest": {
        "type": "object",
        "required": [
          "office_name",
          "alias",
          "pet_id",
          "presence_visible",
          "stats_opt_in",
          "poster_opt_in"
        ],
        "properties": {
          "office_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[^\\r\\n]+$"
          },
          "alias": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32,
            "pattern": "^[^@\\r\\n]+$"
          },
          "pet_id": {
            "enum": [
              "bloop",
              "buzz-bit",
              "comet-lop",
              "gizmo-kernel",
              "kiki-koala",
              "momo-mug",
              "moss-shell",
              "nimbus-noodle",
              "nori-nibble",
              "olli-orbit",
              "pico-patch",
              "pogo-ping",
              "rumi-relay",
              "sora-shiba",
              "taro-tinker"
            ]
          },
          "presence_visible": {
            "type": "boolean"
          },
          "stats_opt_in": {
            "type": "boolean"
          },
          "poster_opt_in": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "OnboardingOfficeCreated": {
        "description": "Private Office topology created and one five-minute pairing issued.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/PrivateNoStore"
          },
          "Referrer-Policy": {
            "$ref": "#/components/headers/NoReferrer"
          },
          "X-Content-Type-Options": {
            "$ref": "#/components/headers/NoSniff"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateOfficeOnboardingResponse"
            }
          }
        }
      },
...
      "OnboardingBadRequest": {
        "description": "The request is malformed, contains extra or missing fields, or violates an onboarding input boundary.",
        "x-problem-codes": [
          "invalid_request"
        ],
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/PrivateNoStore"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/OnboardingProblem"
                },
                {
                  "type": "object",
                  "properties": {
                    "type": {
                      "const": "https://workbuddy-buddy.invalid/problems/invalid_request"
                    },
                    "title": {
                      "const": "Invalid request"
                    },
                    "status": {
                      "const": 400
                    },
                    "code": {
                      "const": "invalid_request"
                    }
                  },
                  "not": {
                    "properties": {
                      "retry_after_seconds": {}
                    },
                    "required": [
                      "retry_after_seconds"
                    ]
                  }
                }
              ]
            }
          }
        }
      },
...
      "OnboardingRateLimited": {
        "description": "The per-IP onboarding window or pairing attempt budget is exhausted.",
        "x-problem-codes": [
          "rate_limited"
        ],
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/PrivateNoStore"
          },
          "Retry-After": {
            "description": "Whole seconds until this request may be retried.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 300
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/OnboardingProblem"
                },
                {
                  "type": "object",
                  "required": [
                    "retry_after_seconds"
                  ],
                  "properties": {
                    "type": {
                      "const": "https://workbuddy-buddy.invalid/problems/rate_limited"
                    },
                    "title": {
                      "const": "Rate limit exceeded"
                    },
                    "status": {
                      "const": 429
                    },
                    "code": {
                      "const": "rate_limited"
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "EdgeReportAccepted": {
        "description": "The report was durably accepted, or was an exact replay from the still-current boot whose original acknowledgement is returned without side effects. A replay from a fenced boot, sequence gaps, partial overlaps, and same-sequence different-content conflicts use a `409` Problem response. Presence is anchored to the authoritative `server_received_at` in the acknowledgement; `idle_stage` remains server-derived.",
        "headers": {
          "X-Min-Supported-Protocol": {
            "$ref": "#/components/headers/MinSupportedProtocol"
          },
          "X-Supported-Protocols": {
            "$ref": "#/components/headers/SupportedProtocols"
          },
          "X-Server-Received-At": {
            "$ref": "#/components/headers/ServerReceivedAt"
          },
          "X-Protocol-Version": {
            "$ref": "#/components/headers/CurrentProtocol"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "./schemas/edge-report-ack.v1.schema.json"
            }
          }
        }
      },
      "BadEdgeReport": {
        "description": "Malformed envelope, invalid canonical form, or invalid sequence layout.",
        "headers": {
          "X-Min-Supported-Protocol": {
            "$ref": "#/components/headers/MinSupportedProtocol"
          },
          "X-Supported-Protocols": {
            "$ref": "#/components/headers/SupportedProtocols"
          },
          "X-Protocol-Version": {
            "$ref": "#/components/headers/CurrentProtocol"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "./schemas/problem.v1.schema.json"
            }
          }
        }
      },
      "UnauthenticatedEdge": {
        "description": "Device key is unknown or revoked, or the signature is invalid.",
        "headers": {
          "X-Min-Supported-Protocol": {
            "$ref": "#/components/headers/MinSupportedProtocol"
          },
          "X-Supported-Protocols": {
            "$ref": "#/components/headers/SupportedProtocols"
          },
          "X-Protocol-Version": {
            "$ref": "#/components/headers/CurrentProtocol"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "./schemas/problem.v1.schema.json"
            }
          }
        }
      },
      "EdgeFenceConflict": {
        "description": "The report uses a fenced boot, fails the `previous_boot_id`\ncompare-and-swap, reuses a sequence with different content, starts after an\nunexplained sequence gap, or conflicts with the active reporting instance.\nRetrying unchanged content is safe; changing `observed_at`, `sent_at`, or\nsignature does not make a conflicting sequence acceptable.\n",
        "x-problem-codes": [
          "stale_boot",
          "sequence_gap",
          "sequence_overlap",
          "sequence_conflict"
        ],
        "headers": {
          "X-Min-Supported-Protocol": {
            "$ref": "#/components/headers/MinSupportedProtocol"
          },
          "X-Supported-Protocols": {
            "$ref": "#/components/headers/SupportedProtocols"
          },
          "X-Protocol-Version": {
            "$ref": "#/components/headers/CurrentProtocol"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "./schemas/problem.v1.schema.json"
            }
          }
        }
      },
      "EdgeBatchTooLarge": {
        "description": "The bounded event count or encoded request size is exceeded.",
        "headers": {
          "X-Min-Supported-Protocol": {
            "$ref": "#/components/headers/MinSupportedProtocol"
          },
          "X-Supported-Protocols": {
            "$ref": "#/components/headers/SupportedProtocols"
          },
          "X-Protocol-Version": {
            "$ref": "#/components/headers/CurrentProtocol"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "./schemas/problem.v1.schema.json"
            }
          }
        }
      },
      "UnprocessableEdgeReport": {
        "description": "The envelope is structurally valid but violates non-conflict protocol\nsemantics, for example `first_sequence` not matching the first event, a new\nboot whose first event is not a state snapshot, or a display/activity\ncombination forbidden by the declared protocol version. Intra-batch gaps and\noverlaps are sequence conflicts and use `409`, not `422`.\n",
        "x-problem-codes": [
          "report_semantics_invalid"
        ],
        "headers": {
          "X-Min-Supported-Protocol": {
            "$ref": "#/components/headers/MinSupportedProtocol"
          },
          "X-Supported-Protocols": {
            "$ref": "#/components/headers/SupportedProtocols"
          },
          "X-Protocol-Version": {
            "$ref": "#/components/headers/CurrentProtocol"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "./schemas/problem.v1.schema.json"
            }
          }
        }
      },
      "UnsupportedProtocol": {
        "description": "The Edge protocol is older than `X-Min-Supported-Protocol` or newer than\nthe server's current protocol. The local desktop pet must keep running;\nonly Control Plane reporting is unavailable until compatibility is\nrestored.\n",
        "headers": {
          "X-Min-Supported-Protocol": {
            "$ref": "#/components/headers/MinSupportedProtocol"
          },
          "X-Supported-Protocols": {
            "$ref": "#/components/headers/SupportedProtocols"
          },
          "X-Upgrade-URL": {
            "$ref": "#/components/headers/UpgradeLink"
          },
          "X-Protocol-Version": {
            "$ref": "#/components/headers/CurrentProtocol"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "./schemas/problem.v1.schema.json"
            }
          }
        }
      },
...
      "EdgeRateLimited": {
        "description": "The signed Edge report was not processed because this enrolled device or source exceeded its reporting budget. No event, lease, receipt, sequence, or scoring state is advanced.",
        "x-problem-codes": [
          "rate_limited"
        ],
        "headers": {
          "Retry-After": {
            "description": "Whole seconds until this Edge should retry the unchanged report.",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 86400
            }
          },
          "X-Min-Supported-Protocol": {
            "$ref": "#/components/headers/MinSupportedProtocol"
          },
          "X-Supported-Protocols": {
            "$ref": "#/components/headers/SupportedProtocols"
          },
          "X-Protocol-Version": {
            "$ref": "#/components/headers/CurrentProtocol"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "./schemas/problem.v1.schema.json"
                },
                {
                  "type": "object",
                  "properties": {
                    "type": {
                      "const": "https://workbuddy-buddy.invalid/problems/rate_limited"
                    },
                    "title": {
                      "const": "Rate limit exceeded"
                    },
                    "status": {
                      "const": 429
                    },
                    "code": {
                      "const": "rate_limited"
                    },
                    "retry_after_seconds": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 86400
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
      "TemporarilyUnavailable": {
        "description": "Control Plane is temporarily unable to serve the request.",
        "headers": {
          "Retry-After": {
            "description": "Seconds until the caller should retry.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "./schemas/problem.v1.schema.json"
            }
          }
        }
      }
    }
  }
}